Add files via upload
Ruby.rb is the slide file. I'm working the kinks out of the left alignment and above and below spacing.
The window.rb is a file for examples of simple ruby code. Will keep adding to it. Let me know what kind of simple code examples to add. Idea is to open both files in RubyMine and switch back and forth. Both can run at the same time without losing your spot in the slide show.